[swappable] Removed abstract admin classes
