+++ tests/openwisp2/sample_firmware_upgrader/admin.py
@@ -1,26 +1,5 @@
-from django.contrib import admin
-from openwisp_firmware_upgrader.base.admin import (
-    AbstractBatchUpgradeOperationAdmin,
-    AbstractBuildAdmin,
-    AbstractCategoryAdmin,
+from openwisp_firmware_upgrader.admin import (  # noqa
+    BatchUpgradeOperationAdmin,
+    BuildAdmin,
+    CategoryAdmin,
 )
-from swapper import load_model
-
-BatchUpgradeOperation = load_model('firmware_upgrader', 'BatchUpgradeOperation')
-Build = load_model('firmware_upgrader', 'Build')
-Category = load_model('firmware_upgrader', 'Category')
-
-
-@admin.register(Category)
-class CategoryAdmin(AbstractCategoryAdmin):
-    pass
-
-
-@admin.register(Build)
-class BuildAdmin(AbstractBuildAdmin):
-    pass
-
-
-@admin.register(BatchUpgradeOperation)
-class BatchUpgradeOperationAdmin(AbstractBatchUpgradeOperationAdmin):
-    pass
